LayoutTests/fast/events/wheelevent-basic-expected.txt

TOP TOP TOP TOP TOP TOP TOP TOP TOP TOP TOP TOP TOP TOP
Scroll mouse wheel over here
Scroll mouse wheel over here
Scroll mouse wheel over here
Scroll mouse wheel over here
Scroll mouse wheel over here
Scroll mouse wheel over here
END END END END END END END END END END END END END END
Tests the basic functionality of the standard wheel event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WheelEvent.prototype.__proto__ is MouseEvent.prototype
PASS WheelEvent.DOM_DELTA_PIXEL is 0x00
PASS WheelEvent.DOM_DELTA_LINE is 0x01
PASS WheelEvent.DOM_DELTA_PAGE is 0x02
PASS window.onwheel is null
PASS document.onwheel is null
PASS testDiv.onwheel is null
PASS testEvent.__proto__ is WheelEvent.prototype
PASS testEvent.__proto__.__proto__ is MouseEvent.prototype
PASS testEvent.deltaZ is 0
PASS testEvent.deltaMode is WheelEvent.DOM_DELTA_PIXEL
PASS deltaX > 0 is true
PASS deltaX is expectedDeltaX
PASS deltaY > 0 is true
PASS deltaY is expectedDeltaY
PASS successfullyParsed is true

TEST COMPLETE
